Rename parser files
