+++ Parser/pegen.c
@@ -3,7 +3,7 @@
 #include "tokenizer.h"
 
 #include "pegen.h"
-#include "parse_string.h"
+#include "string_parser.h"
 
 PyObject *
 _PyPegen_new_type_comment(Parser *p, char *s)
